Remove all (float) and replace Math with Mathf.

+++ Runtime/animation/animation.cs
@@ -58,7 +58,7 @@ namespace Unity.UIWidgets.animation {
         }
 
         public Animation<U> drive<U>(Animatable<U> child) {
-            D.assert(this is Animation<double>);
+            D.assert(this is Animation<float>);
             return child.animate(this as Animation<float>);
         }
     }
